+++ packages/xbell/src/types/config.ts
@@ -1,5 +1,5 @@
 import type { UserConfigExport } from 'vite';
-import type { XBellTestCaseFunction } from './test';
+import type { XBellNodeJSTestCaseFunction } from './test';
 import type { StorageState } from './pw';
 export type XBellBrowserType = 'chromium' | 'firefox' | 'webkit';
 
@@ -47,8 +47,8 @@ export interface XBellConfig {
   maxThreads?: number;
 
   hooks?: {
-    beforeEach?: XBellTestCaseFunction;
-    afterEach?: XBellTestCaseFunction;
+    beforeEach?: XBellNodeJSTestCaseFunction;
+    afterEach?: XBellNodeJSTestCaseFunction;
   }
 
   coverage?: {
@@ -109,8 +109,16 @@ export interface XBellNodeJSCallback {
   sortValue: number; // no used
 }
 
+export interface XBellCommonCallback {
+  callback: (...args: any) => any;
+  filename: string;
+  sortValue: number; // no used
+}
+
+
 export type XBellRuntimeOptions = Partial<{
   browserCallbacks: XBellBrowserCallback[];
   nodejsCallbacks: XBellNodeJSCallback[];
+  commonCallbacks: XBellCommonCallback[];
   args?: object
 }>;