+++ packages/xbell/CHANGELOG.md
@@ -1,5 +1,47 @@
 # xbell
 
+## 0.8.0
+
+### Minor Changes
+
+- 60af676: feat: support test.all for run test in all runtimes
+
+### Patch Changes
+
+- Updated dependencies [60af676]
+  - @xbell/code-stack@0.8.0
+  - @xbell/coverage@0.8.0
+  - @xbell/reporter@0.8.0
+  - @xbell/snapshot@0.8.0
+  - @xbell/assert@0.8.0
+  - @xbell/color@0.8.0
+
+## 0.7.2
+
+### Patch Changes
+
+- a656ab5: add FileChooser
+- Updated dependencies [a656ab5]
+  - @xbell/assert@0.7.2
+  - @xbell/code-stack@0.7.2
+  - @xbell/color@0.7.2
+  - @xbell/coverage@0.7.2
+  - @xbell/reporter@0.7.2
+  - @xbell/snapshot@0.7.2
+
+## 0.7.1
+
+### Patch Changes
+
+- a010901: support wait for file chooser
+- Updated dependencies [a010901]
+  - @xbell/assert@0.7.1
+  - @xbell/code-stack@0.7.1
+  - @xbell/color@0.7.1
+  - @xbell/coverage@0.7.1
+  - @xbell/reporter@0.7.1
+  - @xbell/snapshot@0.7.1
+
 ## 0.7.0
 
 ### Patch Changes